move functions into class
